refactor: Add `AssetManager.copyPathToDir` extension
Add `AssetManager.copyPathToDir` extension method to copy assets files or directories recursively to a target directory. This method handles both files and directories by determining if the path points to a file or a directory.

- If the path is a file, it is copied to the target directory.
- If the path is a directory, it recursively copies its contents to the target directory, creating subdirectories as needed.

An `overwrite` option is available to control whether existing files in the target directory should be overwritten.

Also included are:
- New helper extensions `AssetManager.isFile` and `AssetManager.isDir`
- Corresponding unit tests in `AssetManagerTest.kt` and test assets.