Refactor: Rename `equals` to `equal` for `CSValue` comparisons
This change refactors the infix functions used for comparing `CSValue`
instances with other values or `CSValue` instances. The function names
have been changed from `equals` and `equalsNot` to `equal` and
`equalNot` respectively.

This renaming is done to avoid potential conflicts or confusion with
the standard `equals` method in Java/Kotlin and to provide a more
distinct naming convention for `CSValue` comparisons.