Add coerceTo extension function for Float
This commit introduces a new extension function `coerceTo(range: ClosedFloatingPointRange<Float>)` for the `Float` type.
This function constrains the float value to be within the given `range` by first taking the minimum of the float and the range start, and then the maximum of the result and the range end.

Additionally, the unused import `kotlin.Float.Companion.MAX_VALUE` has been removed.