Ignore DeadSystemException in CSApplication
This commit updates the `CSApplication` class to ignore `DeadSystemException` and `DeadSystemRuntimeException` in the global uncaught exception handler.

A new private extension function `Throwable.isSystemDead()` was added to check if a throwable or its cause is one of these exceptions. The uncaught exception handler now calls this function and logs a warning if it returns true.